requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.3.1
auto-py-to-exe==2.8.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.12.5
cffi==1.14.5
cryptography==3.4.6
distlib==0.3.1
Django==3.1.7
django-rest-knox==4.1.0
djangorestframework==3.12.2
Eel==0.12.4
filelock==3.0.12
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
gunicorn==20.0.4
macholib==1.14
pipenv==2020.11.15
psycopg2-binary==2.8.6
pycparser==2.20
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
virtualenv==20.4.3
virtualenv-clone==0.5.4
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.3.0
